backend/requirements.txt
# Ultra-minimal requirements - no Pydantic
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
